requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
llvmlite==0.28.0
matplotlib==3.0.3
mpmath==1.1.0
numba==0.43.1
numpy==1.16.3
Pillow==6.0.0
pyparsing==2.4.0
PyQt5==5.12.2
PyQt5-sip==4.19.17
python-dateutil==2.8.0
scipy==1.2.1
six==1.12.0
sympy==1.4